Rename files to .js - Fix typo with
