+++ README.md
@@ -111,7 +111,7 @@ npm run develop # or gatsby develop
 
 ## Configuration
 
-Edit the export object in `data/SiteConfig`:
+Edit the export object in `data/siteConfig`:
 
 ```js
 module.exports = {
