--- src/components/About/About.css
@@ -1,7 +1,0 @@
-.about {
-    display: flex;
-    justify-content: center;
-    align-content: center;
-    align-items: center;
-    min-height: 300px;
-}
