+++ src/components/SocialLinks/SocialLinks.js
@@ -18,7 +18,7 @@ import {
   RedditIcon
 } from "react-share";
 import urljoin from "url-join";
-import config from "../../../data/SiteConfig";
+import config from "../../../data/siteConfig";
 import "./SocialLinks.css";
 
 class SocialLinks extends Component {
