+++ src/layout/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Helmet from "react-helmet";
-import config from "../../data/SiteConfig";
+import config from "../../data/siteConfig";
 import "./index.css";
 
 export default class MainLayout extends React.Component {
