--- src/pages/about.jsx
@@ -1,20 +1,0 @@
-import React, { Component } from "react";
-import Helmet from "react-helmet";
-import Layout from "../layout";
-import About from "../components/About/About";
-import config from "../../data/SiteConfig";
-
-class AboutPage extends Component {
-  render() {
-    return (
-      <Layout>
-        <div className="about-container">
-          <Helmet title={`About | ${config.siteTitle}`} />
-          <About />
-        </div>
-      </Layout>
-    );
-  }
-}
-
-export default AboutPage;
