+++ src/templates/category.js
@@ -3,16 +3,18 @@ import Helmet from "react-helmet";
 import { graphql } from "gatsby";
 import Layout from "../layout";
 import PostListing from "../components/PostListing/PostListing";
-import config from "../../data/SiteConfig";
+import config from "../../data/siteConfig";
 
-export default class TagTemplate extends React.Component {
+export default class CategoryTemplate extends React.Component {
   render() {
-    const { tag } = this.props.pageContext;
+    const { category } = this.props.pageContext;
     const postEdges = this.props.data.allMarkdownRemark.edges;
     return (
       <Layout>
-        <div className="tag-container">
-          <Helmet title={`Posts tagged as "${tag}" | ${config.siteTitle}`} />
+        <div className="category-container">
+          <Helmet
+            title={`Posts in category "${category}" | ${config.siteTitle}`}
+          />
           <PostListing postEdges={postEdges} />
         </div>
       </Layout>
@@ -22,11 +24,11 @@ export default class TagTemplate extends React.Component {
 
 /* eslint no-undef: "off" */
 export const pageQuery = graphql`
-  query TagPage($tag: String) {
+  query CategoryPage($category: String) {
     allMarkdownRemark(
       limit: 1000
       sort: { fields: [fields___date], order: DESC }
-      filter: { frontmatter: { tags: { in: [$tag] } } }
+      filter: { frontmatter: { category: { eq: $category } } }
     ) {
       totalCount
       edges {
